# sdk-server/docs/swagger.yaml
definitions:
  base.PageResp:
    properties:
      currentPage:
        description: 当前第几页
        type: integer
      list:
        description: 数据列表
      pageSize:
        description: 分页大小
        type: integer
      total:
        description: 总条数
        type: integer
    type: object
  base.ReqId:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
  base.ReqIds:
    properties:
      ids:
        description: 多id
        items:
          type: integer
        type: array
    type: object
  base.Resp:
    properties:
      code:
        description: 返回码
        type: integer
      data:
        description: 数据
      msg:
        description: 消息
        type: string
      reqId:
        description: '`json:"请求id"`'
        type: string
    type: object
  brosdk.EnvInfo:
    properties:
      bridgeProxy:
        description: 桥代理配置，格式为：socks5://user:pwd@ipaddr:6666
        type: string
      cookie:
        description: cookie
        type: string
      customerId:
        description: 三方用户id
        type: string
      envId:
        description: envid 更新时必填
        type: string
      envName:
        description: 环境名称
        type: string
      finger:
        $ref: '#/definitions/brosdk.Finger'
      ipChannel:
        description: IP监测渠道  海外代理：ip2location，国内代理：ipdata
        type: string
      proxy:
        description: 代理配置，格式为：socks5://user:pwd@ipaddr:6666
        type: string
      region:
        description: 国家代号，当无法获取代理配置时，传此参数生成对应区域ip，否则获取客户端ip
        type: string
      remark:
        description: 备注
        type: string
      serial:
        description: 环境序号
        type: string
    type: object
  brosdk.Finger:
    properties:
      audioContext:
        description: AudioContext 1隐身 2真实
        type: integer
      battery:
        description: 电池 1真实 2噪声
        type: integer
      bluetooth:
        description: 蓝牙 1开启 2关闭(默认)
        type: integer
      canvas:
        description: 浏览器canvas指纹开关 0:倾向一致性 1:关闭 2:倾向随机性
        type: integer
      clientRects:
        description: ClientRects 1隐身 2真实 目前同fontfinger值
        type: integer
      cpu:
        description: CPU核心数量 不传会自动生成
        type: integer
      deviceName:
        description: 计算机名 不传会自动生成
        type: string
      doNotTrack:
        description: 请勿跟踪浏览器设置   1不启用 2启用 3默认
        type: integer
      dpi:
        description: 平面分辨率 空自动生成
        type: string
      enableGc:
        description: 是否开启垃圾回收 1开启 2关闭
        type: integer
      enableNotice:
        description: 网页通知 1开启，2关闭
        type: integer
      enableOpen:
        description: 多开设置 1开启，2关闭
        type: integer
      enableOpenNumber:
        description: 多开人数设置
        type: integer
      enablePic:
        description: 禁止加载图片 1开启，2关闭
        type: integer
      enableScanPort:
        description: 端口扫描防护 1开启(默认) 2关闭
        type: integer
      enableSound:
        description: 禁止播放声音 1开启，2关闭
        type: integer
      enableVideo:
        description: 禁止加载视频 1开启，2关闭
        type: integer
      font:
        allOf:
        - $ref: '#/definitions/brosdk.Font'
        description: 字体列表
      fontFinger:
        description: 字体指纹 1隐身 2真实
        type: integer
      gcTime:
        description: 垃圾回收时间,当enableGc为1时必填 1-5
        type: integer
      geographic:
        allOf:
        - $ref: '#/definitions/brosdk.Geographic'
        description: 地理位置 （默认使用IP定位 动态代理IP不支持次选项为禁止）
      hardware:
        description: 硬件加速 1开启（默认） 2关闭
        type: integer
      kernel:
        description: 内核
        type: string
      kernelVersion:
        description: 内核版本
        type: string
      language:
        description: 浏览器的语言 不传会根据代理IP地址自动生成 详细看支持的语言详细列表(如果是使用动态IP自动生成为中文)
        items:
          type: string
        type: array
      mac:
        description: MAC地址 不传会自动生成
        type: string
      mediaDevice:
        description: 媒体设备开关，1：关闭（每个浏览器使用当前电脑默认的媒体设备id，真实）2：启用（使用相匹配的值代替您真实的媒体设备ID，噪声）（默认）
        type: integer
      mem:
        description: 内存参数 不传会自动生成
        type: number
      nextsystem:
        allOf:
        - $ref: '#/definitions/brosdk.NextSystem'
        description: （可选不使用设置为空）指纹生成的时候回根据System 或者 Nextsystem去生成，如果给了多项会从这些当中随机选择生成
      picSize:
        description: 图片大小
        type: string
      scanPort:
        description: 白名单 0~65535 关闭状态不写 当EnableScanPort是1时这里为空会自动生成本地端口
        type: string
      speechVoices:
        description: SpeechVoices指纹，1：每个浏览器使用当前电脑默认的SpeechVoices,真实 2：添加相应的噪音，同一电脑上为每个浏览器生成不同的SpeechVoices（默认）
        type: integer
      system:
        description: 系统
        type: string
      ua:
        description: 不写根据系统和浏览器版本自动生成
        type: string
      uaVersion:
        description: 浏览器大版本号 不传自动生成 详细查看具体支持的版本号大版本
        type: string
      uilanguage:
        description: 界面语言 空会根据代理IP地址自动生成
        items:
          type: string
        type: array
      webGl:
        description: 浏览器webgl元数据指纹开关 1隐身 2真实（默认）
        type: integer
      webGlInfo:
        description: 浏览器WebGlInfo 1:真实 2:自定义
        type: integer
      webGlRenderer:
        description: 浏览器WebGL渲染 自定义时传值，为空会自动生成，该字段不为空时webGLVendor必传
        type: string
      webGlVendor:
        description: 浏览器WebGL厂商，Windows系统可选值为Google Inc. (NVIDIA)、Google Inc. (AMD)、Google
          Inc. (Intel)，MacOS系统可选值为Google Inc. (ATI Technologies Inc.)、Google Inc.
          (NVIDIA)、Google Inc. (Apple)，Android系统可选值为Qualcomm，IOS系统可选值为Apple Inc 自定义时传值，为空会自动生成
        type: string
      webRTC:
        description: WebRTC 0:禁用,网站会拿不到IP 1:真实,网站会获取真实IP 2:替换,使用代理IP覆盖真实IP
        type: integer
      webRTCIP:
        description: Chrome即时通信组件，支持：proxy 替换 ，使用代理IP覆盖真实IP，代理场景使用 local 真实 ，网站会获取真实IP
          disabled 禁用(默认)，网站会拿不到IP
        type: string
      widowssize:
        description: 浏览器窗口大小  （目前可传可不传已经不用了使用dpi计算得出）
        type: string
      zone:
        description: 时区 不传会根据代理IP地址自动生成 详细查看时区支持的列表(如果是使用动态IP自动生成为北京时间)
        type: string
    type: object
  brosdk.Font:
    properties:
      enable:
        description: 1隐身包含 2使用用户设置
        type: integer
      list:
        description: enable=2开启情况下如果为空系统自动生成
        items:
          type: string
        type: array
    type: object
  brosdk.Geographic:
    properties:
      accuracy:
        description: 精度（米）(当enable等于2且UseIP等于0时使用) 10 - 5000
        type: string
      enable:
        description: 启用1(默认) 询问2 禁止3
        type: integer
      latitude:
        description: 纬度(当enable等于2且UseIP等于0时使用) -90 - 90
        type: string
      longitude:
        description: 经度(当enable等于2且UseIP等于0时使用) -180 - 180
        type: string
      useip:
        description: 1使用ip定位(默认),2使用自定义
        type: integer
    type: object
  brosdk.NextSystem:
    properties:
      Android:
        type: string
      IOS:
        type: string
      Linux:
        type: string
      MacOS:
        type: string
    type: object
...
  dto.BrowserGetPageReq:
    properties:
      envId:
        description: 环境ID
        type: string
      name:
        description: 名称
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      userId:
        description: 用户ID
        type: integer
    type: object
...
  models.Browser:
    properties:
      createdAt:
        description: 创建时间
        type: string
      data:
        description: 数据
        type: string
      envId:
        description: 环境ID
        type: string
      id:
        description: 主键
        type: integer
      name:
        description: 名称
        type: string
      status:
        description: 状态 1 停止 3启动
        type: integer
      updatedAt:
        description: 最后更新时间
        type: string
      userId:
        description: 用户ID
        type: integer
    type: object
...
info:
  contact:
    email: baowk1024@gmail.com
    name: baowk
    url: https://github.com/baowk/dilu
  description: 致力于做一个开发快速，运行稳定的框架
  title: Dilu API
  version: V0.0.1
paths:
...
  /api/app/browser/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/brosdk.EnvInfo'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  type: string
              type: object
      security:
      - Bearer: []
      summary: 创建Browser
      tags:
      - app-browser
  /api/app/browser/del:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/base.ReqIds'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Browser'
              type: object
      security:
      - Bearer: []
      summary: 删除Browser
      tags:
      - app-browser
  /api/app/browser/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/base.ReqId'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Browser'
              type: object
      security:
      - Bearer: []
      summary: 获取Browser
      tags:
      - app-browser
  /api/app/browser/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.BrowserGetPageReq'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/base.PageResp'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/models.Browser'
                        type: array
                    type: object
              type: object
      security:
      - Bearer: []
      summary: Page接口
      tags:
      - app-browser
  /api/app/browser/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/brosdk.EnvInfo'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  type: string
              type: object
      security:
      - Bearer: []
      summary: 更新Browser
      tags:
      - app-browser
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"